Add a minor refactoring
