+++ src/brain/cognition/understanding/text.js
@@ -1,0 +1,7 @@
+import { serializer } from "../../utils";
+
+const extractFeeling = (sentence, cortex) => cortex.run(serializer.encode(sentence))
+
+export {
+  extractFeeling,
+}
