+++ src/brain/knowledge/feelings/index.js
@@ -1,0 +1,5 @@
+import hapiness from './happiness'
+import sadness from './sadness'
+
+
+export const feelings = [...new Set(hapiness, sadness)]
