--- src/brain/knowledge/training-data/index.js
@@ -1,9 +1,0 @@
-const happy = require('./d-happy')
-const sad   = require('./d-sad')
-
-const moods = [
-  ...happy,
-  ...sad
-];
-
-module.exports = moods
